requirements-pinned.txt
Fabric==1.8.3
Flask==0.10.1
Flask-Assets==0.9
Flask-Login==0.2.9
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
cffi==0.8.1
common==0.1.1
docopt==0.4.0
ecdsa==0.11
itsdangerous==0.23
jdcal==1.0
mandrill==1.0.53
mock==1.0.1
nose==1.3.4
openpyxl==2.0.4
paramiko==1.12.3
protobuf==2.5.0
pycparser==2.10
pycrypto==2.6.1
pygal==1.5.0
python-dateutil==2.2
pytz==2013.9
requests==2.3.0
rethinkdb
selenium==2.42.1
six==1.7.2
tzlocal==1.1.1
virtualenv==1.10.1
webassets==0.9
xlrd==0.9.2
yolk==0.4.3
